Réalisation du système de pagination des applications.

+++ src/Gary/src/main/java/ch/heigvd/amt/gary/controllers/AppsListServlet.java
@@ -41,7 +41,7 @@ public class AppsListServlet extends HttpServlet {
                                        1 : Integer.parseInt(request.getParameter("page"));
          int itemsPerPage = request.getParameter("per_page") == null ? 
                                        10 : Integer.parseInt(request.getParameter("per_page"));
-         int numberOfPages = (int)Math.ceil(appsManager.countForAccount((long)request.getSession().getAttribute("id")) / itemsPerPage);
+         int numberOfPages = (int)Math.ceil(appsManager.countForAccount((long)request.getSession().getAttribute("id")) / (double)itemsPerPage);
          
          Object apps = appsManager.getUserApps((long)request.getSession().getAttribute("id"), 
                                                pageNumber,
